Move AsyncImageFileView to Common, which required some more shuffling around of stuff.
