+++ UI/PauseScreen.h
@@ -24,7 +24,11 @@
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/ViewGroup.h"
 #include "UI/MiscScreens.h"
-#include "UI/TextureUtil.h"
+
+enum class PauseScreenMode {
+	MAIN,
+	DISPLAY_SETTINGS,
+};
 
 class GamePauseScreen : public UIDialogScreenWithGameBackground {
 public:
@@ -64,41 +68,5 @@ private:
 	// hack
 	bool finishNextFrame_ = false;
 	Path gamePath_;
-};
-
-// AsyncImageFileView loads a texture from a file, and reloads it as necessary.
-// TODO: Actually make async, doh.
-class AsyncImageFileView : public UI::Clickable {
-public:
-	AsyncImageFileView(const Path &filename, UI::ImageSizeMode sizeMode, UI::LayoutParams *layoutParams = 0);
-	~AsyncImageFileView();
-
-	void GetContentDimensionsBySpec(const UIContext &dc, UI::MeasureSpec horiz, UI::MeasureSpec vert, float &w, float &h) const override;
-	void Draw(UIContext &dc) override;
-	std::string DescribeText() const override { return text_; }
-
-	void DeviceLost() override;
-	void DeviceRestored(Draw::DrawContext *draw) override;
-
-	void SetFilename(const Path &filename);
-	void SetColor(uint32_t color) { color_ = color; }
-	void SetOverlayText(std::string text) { text_ = text; }
-	void SetFixedSize(float fixW, float fixH) { fixedSizeW_ = fixW; fixedSizeH_ = fixH; }
-	void SetCanBeFocused(bool can) { canFocus_ = can; }
-
-	bool CanBeFocused() const override { return canFocus_; }
-
-	const Path &GetFilename() const { return filename_; }
-
-private:
-	bool canFocus_;
-	Path filename_;
-	std::string text_;
-	uint32_t color_;
-	UI::ImageSizeMode sizeMode_;
-
-	std::unique_ptr<ManagedTexture> texture_;
-	bool textureFailed_;
-	float fixedSizeW_;
-	float fixedSizeH_;
+	PauseScreenMode mode_ = PauseScreenMode::MAIN;
 };
