+++ UI/UI.vcxproj.filters
@@ -67,7 +67,6 @@
     <ClCompile Include="RemoteISOScreen.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
-    <ClCompile Include="TextureUtil.cpp" />
     <ClCompile Include="DiscordIntegration.cpp" />
     <ClCompile Include="GPUDriverTestScreen.cpp">
       <Filter>Screens</Filter>
@@ -81,6 +80,7 @@
     <ClCompile Include="MemStickScreen.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
+    <ClCompile Include="Theme.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GameInfoCache.h" />
@@ -149,7 +149,6 @@
     <ClInclude Include="RemoteISOScreen.h">
       <Filter>Screens</Filter>
     </ClInclude>
-    <ClInclude Include="TextureUtil.h" />
     <ClInclude Include="DiscordIntegration.h" />
     <ClInclude Include="GPUDriverTestScreen.h">
       <Filter>Screens</Filter>
@@ -163,6 +162,7 @@
     <ClInclude Include="MemStickScreen.h">
       <Filter>Screens</Filter>
     </ClInclude>
+    <ClInclude Include="Theme.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Screens">
