+++ UWP/CommonUWP/CommonUWP.vcxproj.filters
@@ -405,6 +405,12 @@
     <ClCompile Include="..\..\Common\GPU\Vulkan\VulkanLoader.cpp">
       <Filter>GPU\Vulkan</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Common\UI\AsyncImageFileView.cpp">
+      <Filter>UI</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Common\Render\ManagedTexture.cpp">
+      <Filter>Render</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="targetver.h" />
@@ -747,6 +753,12 @@
     </ClInclude>
     <ClInclude Include="..\..\Common\GPU\Vulkan\VulkanLoader.h">
       <Filter>GPU\Vulkan</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Common\UI\AsyncImageFileView.h">
+      <Filter>UI</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\Common\Render\ManagedTexture.h">
+      <Filter>Render</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
