+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -28,11 +28,12 @@
     <ClCompile Include="..\..\UI\ReportScreen.cpp" />
     <ClCompile Include="..\..\UI\SavedataScreen.cpp" />
     <ClCompile Include="..\..\UI\Store.cpp" />
-    <ClCompile Include="..\..\UI\TextureUtil.cpp" />
     <ClCompile Include="..\..\UI\TiltAnalogSettingsScreen.cpp" />
     <ClCompile Include="..\..\UI\TiltEventProcessor.cpp" />
     <ClCompile Include="..\..\UI\TouchControlLayoutScreen.cpp" />
     <ClCompile Include="..\..\UI\TouchControlVisibilityScreen.cpp" />
+    <ClCompile Include="..\..\UI\ChatScreen.cpp" />
+    <ClCompile Include="..\..\UI\Theme.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -63,10 +64,11 @@
     <ClInclude Include="..\..\UI\ReportScreen.h" />
     <ClInclude Include="..\..\UI\SavedataScreen.h" />
     <ClInclude Include="..\..\UI\Store.h" />
-    <ClInclude Include="..\..\UI\TextureUtil.h" />
     <ClInclude Include="..\..\UI\TiltAnalogSettingsScreen.h" />
     <ClInclude Include="..\..\UI\TiltEventProcessor.h" />
     <ClInclude Include="..\..\UI\TouchControlLayoutScreen.h" />
     <ClInclude Include="..\..\UI\TouchControlVisibilityScreen.h" />
+    <ClInclude Include="..\..\UI\ChatScreen.h" />
+    <ClInclude Include="..\..\UI\Theme.h" />
   </ItemGroup>
 </Project>
